refactor: extract parseInstallCmd to lib/install-cmd-parser.js
